backend/travel_planner/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
sqlparse==0.5.0
python-dotenv==1.0.0
google-generativeai==0.3.1